修复部分内存问题

+++ test/Class/Base/View/LBCell.m
@@ -59,7 +59,7 @@
 
 - (void)imageViewClick:(UITapGestureRecognizer *)tap {
     if (_callBack) {
-        _callBack(self.model,tap.view.tag);
+        _callBack(self.model,(int)tap.view.tag);
     }
 }
